Add dry run

+++ BayesOpt4dftu/main.py
@@ -12,6 +12,7 @@ def main():
     out_file_name = vasp_env_params.get('out_file_name', 'vasp.out')
     # Define the path direct to the VASP pseudopotential.
     vasp_pp_path = vasp_env_params.get('vasp_pp_path', '/home/maituoy/pp_vasp/')
+    dry_run = vasp_env_params.get('dry_run', False)
 
     bo_params = data['bo']
     k = float(bo_params.get('kappa', 5))
@@ -27,8 +28,6 @@ def main():
 
     os.environ['VASP_PP_PATH'] = vasp_pp_path
 
-    # calculate(command=vasp_run_command, outfilename=out_file_name, method='hse', import_kpath = import_kpath)
-
     header = []
     for i, u in enumerate(which_u):
         header.append('U_ele_%s' % str(i + 1))
@@ -39,24 +38,36 @@ def main():
     with open('./u_tmp.txt', 'w+') as f:
         f.write('%s band_gap delta_band \n' % (' '.join(header)))
 
-    obj = 0
-    for i in range(iteration):
-        calculate(command=vasp_run_command, outfilename=out_file_name, method='dftu', import_kpath=import_kpath)
-        db = delta_band(bandrange=br, path='./')
-        db.deltaBand()
+    if dry_run:
+        calculate(command=vasp_run_command, outfilename=out_file_name, method='hse', import_kpath=import_kpath,
+                  is_dry=True)
+        calculate(command=vasp_run_command, outfilename=out_file_name, method='dftu', import_kpath=import_kpath,
+                  is_dry=True)
+        print("Dry run executed. No actual calculations were performed. ")
+        print("Review the input files before proceeding.")
 
-        bayesian_opt = bayesOpt_DFTU(path='./', opt_u_index=which_u, u_range=urange, kappa=k, a1=a1, a2=a2,
-                                     elements=elements)
-        obj_next = bayesian_opt.bo()
-        if abs(obj_next - obj) <= threshold:
-            print("Optimization has been finished!")
-            break
-        obj = obj_next
+    else:
+        calculate(command=vasp_run_command, outfilename=out_file_name, method='hse', import_kpath=import_kpath,
+                  is_dry=False)
+        obj = 0
+        for i in range(iteration):
+            calculate(command=vasp_run_command, outfilename=out_file_name, method='dftu', import_kpath=import_kpath,
+                      is_dry=False)
+            db = delta_band(bandrange=br, path='./')
+            db.deltaBand()
 
-    bayesian_opt.plot()
-    print(bayesian_opt.optimal)
+            bayesian_opt = bayesOpt_DFTU(path='./', opt_u_index=which_u, u_range=urange, kappa=k, a1=a1, a2=a2,
+                                         elements=elements)
+            obj_next = bayesian_opt.bo()
+            if abs(obj_next - obj) <= threshold:
+                print("Optimization has been finished!")
+                break
+            obj = obj_next
 
-    os.system('mv ./u_tmp.txt ./u_kappa_%s_a1_%s_a2_%s.txt' % (k, a1, a2))
+        bayesian_opt.plot()
+        print(bayesian_opt.optimal)
+
+        os.system('mv ./u_tmp.txt ./u_kappa_%s_a1_%s_a2_%s.txt' % (k, a1, a2))
 
 
 if __name__ == "__main__":
